refactor: 도메인과 UI 분리
