+++ app/src/main/java/woowacourse/payments/domain/OwnerName.kt
@@ -5,11 +5,11 @@ import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 import woowacourse.payments.domain.exception.OwnerNameException
 
-@Parcelize
+
 @JvmInline
 value class OwnerName(
     val value: String?,
-) : Parcelable {
+){
     init {
         if (value != null) {
             require(value.length <= OWNER_NAME_MAX_LENGTH) { OwnerNameException.OwnerNameMaxLengthException.message }
