+++ app/src/main/java/woowacourse/payments/ui/model/NewCardMode.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.newcard.uiModel
+package woowacourse.payments.ui.model
 
 sealed interface NewCardMode {
     data object CreateMode : NewCardMode
